leetcode/0495-teemoAttacking_test.go
package leetcode

import "testing"

func Test_findPoisonedDuration(t *testing.T) {
	type args struct {
		timeSeries []int
		duration   int
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{"0", args{timeSeries: []int{1, 4}, duration: 2}, 4},
		{"1", args{timeSeries: []int{1, 2}, duration: 2}, 3},
		{"2", args{timeSeries: []int{}, duration: 2}, 0},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := findPoisonedDuration(tt.args.timeSeries, tt.args.duration); got != tt.want {
				t.Errorf("findPoisonedDuration() = %v, want %v", got, tt.want)
			}
		})
	}
}
